Phase 5: Decompose IncrementalAnalyzer into focused modules
- Extract change-category handlers into change_handler.py
  (compile_commands.json, header, source, removed-file handling).
- Extract symbol merge logic into symbol_merger.py
  (definition-wins deduplication, batch merging into symbol store).
- Extract compile-argument precomputation into compile_args_resolver.py.
- Extract process-pool orchestration into worker_orchestrator.py
  (executor creation, task submission, result processing, progress reporting).
- Keep IncrementalAnalyzer as a coordinator with thin public/private wrappers.
- Update test patch target for ProcessPoolExecutor.
- Full test suite passes: 1432 passed, 17 skipped, 4 xpassed.
